+++ handlers.py
@@ -1,0 +1,80 @@
+import logging
+from telegram import Update, InlineKeyboardMarkup, InlineKeyboardButton
+
+
+import constants
+import helpers
+
+
+async def handle_back_to_start(update: Update):
+    logging.info(f"back_to_start triggered by {helpers.get_user(update)}")
+    await update.callback_query.edit_message_text(
+        text=constants.club_description,
+        reply_markup=helpers.main_menu()
+    )
+
+
+async def handle_how_to_join(update: Update):
+    logging.info(f"how_to_join triggered by {helpers.get_user(update)}")
+    await update.callback_query.edit_message_text(
+        text=constants.how_to_join_description,
+        reply_markup=helpers.back_menu())
+    return
+
+
+async def handle_ddia(update: Update):
+    logging.info(f"ddia triggered by {helpers.get_user(update)}")
+    button_list = [
+        InlineKeyboardButton("Хочу сделать презентацию!", callback_data="how_to_present"),
+        InlineKeyboardButton("Назад", callback_data="back"),
+    ]
+    menu = [button_list[i:i + 1] for i in range(0, len(button_list), 1)]
+    await update.callback_query.edit_message_text(
+        text=constants.ddia_description,
+        reply_markup=InlineKeyboardMarkup(menu),
+        parse_mode="HTML")
+    return
+
+
+async def handle_back_to_ddia(update: Update):
+    logging.info(f"back_to_ddia triggered by {helpers.get_user(update)}")
+    await handle_ddia(update)
+
+
+async def handle_sre_book(update: Update):
+    logging.info(f"sre_book triggered by {helpers.get_user(update)}")
+    button_list = [
+        InlineKeyboardButton("Назад", callback_data="back"),
+    ]
+    menu = [button_list[i:i + 1] for i in range(0, len(button_list), 1)]
+    await update.callback_query.edit_message_text(
+        text=constants.sre_book_description,
+        reply_markup=InlineKeyboardMarkup(menu),
+        parse_mode="HTML")
+    return
+
+
+async def handle_mock_leetcode(update: Update):
+    logging.info(f"mock_leetcode triggered by {helpers.get_user(update)}")
+    button_list = [
+        InlineKeyboardButton("Назад", callback_data="back"),
+    ]
+    menu = [button_list[i:i + 1] for i in range(0, len(button_list), 1)]
+    await update.callback_query.edit_message_text(
+        text=constants.mock_leetcode_description,
+        reply_markup=InlineKeyboardMarkup(menu),
+        parse_mode="HTML")
+    return
+
+
+async def handle_how_to_present(update: Update):
+    logging.info(f"how_to_present triggered by {helpers.get_user(update)}")
+    button_list = [
+        InlineKeyboardButton("Назад", callback_data="back_to_ddia"),
+    ]
+    menu = [button_list[i:i + 1] for i in range(0, len(button_list), 1)]
+    await update.callback_query.edit_message_text(
+        text=constants.how_to_present_description,
+        reply_markup=InlineKeyboardMarkup(menu),
+        parse_mode="HTML")
+    return
